[bot-automerge] verilator v5.052

+++ recipe/build.sh
@@ -30,7 +30,9 @@ if [[ "$(uname)" == "Darwin" ]]; then
     # static linking strategy of make_protect_lib example
     # doesn't work with clang on osx-64. End up with
     #   ld: library not found for -lcrt0.o
-    rm -rf examples/make_protect_lib
+    if [[ "$(uname -m)" == "x86_64" ]]; then
+        rm -rf examples/make_protect_lib
+    fi
 
     # work around https://github.com/verilator/verilator/issues/3283
     export LDFLAGS="$LDFLAGS -undefined dynamic_lookup"
@@ -40,5 +42,3 @@ if [[ "$CONDA_BUILD_CROSS_COMPILATION" != "1" ]]; then
     # don't make test when cross compiling. At least until we figure out how to run in an emulator. qemu-static?
     make -j$CPU_COUNT test
 fi
-
-
